Avoid some warnings

+++ mediancut.c
@@ -362,7 +362,7 @@ colormap *mediancut(histogram *hist, const float min_opaque_val, unsigned int ne
         // hist_item_sort_halfvar sorts and sums lowervar at the same time
         // returns item to break at …minus one, which does smell like an off-by-one error.
         hist_item *break_p = hist_item_sort_halfvar(&achv[indx], clrs, &lowervar, halfvar);
-        unsigned int break_at = MIN(clrs-1, break_p - &achv[indx] + 1);
+        unsigned int break_at = MIN(clrs-1, (unsigned int)(break_p - &achv[indx] + 1));
 
         /*
          ** Split the box.
