+++ rwpng.c
@@ -69,7 +69,7 @@ static void user_read_data(png_structp png_ptr, png_bytep data, png_size_t lengt
 
 png_bytepp rwpng_create_row_pointers(png_infop info_ptr, png_structp png_ptr, unsigned char *base, unsigned int height, unsigned int rowbytes)
 {
-    if (!rowbytes) rowbytes = png_get_rowbytes(png_ptr, info_ptr);
+    if (!rowbytes) rowbytes = (unsigned int)png_get_rowbytes(png_ptr, info_ptr);
 
     png_bytepp row_pointers = malloc(height * sizeof(row_pointers[0]));
     for(unsigned int row = 0;  row < height;  ++row) {
